+++ CHANGELOG.md
@@ -2,7 +2,11 @@
 
 ## 0.0.2
 
-- Support re-opening any widget proxy content with `#content {}`
+- Support re-opening any widget proxy content with `#content {}` block to add more nested widgets inside
+- Paint `drawing_area` Cairo shapes declaratively
+- samples/hello/hello_drawing_area.rb (declarative cairo drawing)
+- samples/hello/hello_drawing_area_manual.rb (use cairo manually with imperative syntax)
+- samples/elaborate/widget_gallery.rb
 
 ## 0.0.1